% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ptv-plot.R
\name{plot.parsimonious}
\alias{plot.parsimonious}
\title{A function plotting the estimated parameter path in a parsimonious class object.}
\usage{
\method{plot}{parsimonious}(x, ...)
}
\arguments{
\item{x}{An object of class parsimonious.}

\item{...}{unused.}
}
\description{
A function plotting the estimated parameter path in a parsimonious class object.
}
\examples{
\dontrun{
data(economics, package="ggplot2")
psr <- tail(economics$psavert,-1) # personal saving rate.
lur <- head(log(economics$unemploy/economics$pop),-1) # lagged log unemployment rate.
ptvmod <- ptvfit(y=psr,x=lur)
plot(ptvmod)
}
}

