% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ptv-print-summary.R
\name{print.parsimonious}
\alias{print.parsimonious}
\title{The print and summary method for parsimonious class objects.}
\usage{
\method{print}{parsimonious}(x, ...)
}
\arguments{
\item{x}{An object of class parsimonious.}

\item{...}{unused.}
}
\description{
The print and summary method for parsimonious class objects.
}

