% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ptv-est.R
\name{ptvfit}
\alias{ptvfit}
\title{A function to estimate a model with parsimoniously time varying parameters.}
\usage{
ptvfit(formula, data, wada = NULL, alpha = 1, lambda = NULL,
  peninit = TRUE, postest = TRUE, ptvcste = FALSE, pmax = NULL, ...)
}
\arguments{
\item{formula}{an object of class formula. An intercept should always be included in the model, do not explicitely remove the intercept from the formula.}

\item{data}{an optional data frame.}

\item{alpha}{the mixing parameter of elastic net. Default 1 (Lasso), set to 0 for ridge regression.}

\item{lambda}{user-defined lambda sequence.}

\item{peninit}{boolean: should the initial value of the parameters be penalized or only the increments in the random walk?}

\item{postest}{boolean: should the post Lasso OLS estimation be performed?}

\item{ptvcste}{boolean: should the intercept be parsimoniously time varying?}

\item{...}{unused.}

\item{weights}{an optional vector of weights (of length r(T-r+1)) used to compute an adaptive lasso. If a weight is equal to zero the variable is automatically excluded from the model.}

\item{pmax:}{maximum number of non-zero parameters to estimate.}
}
\value{
An object of the parsimonious class.
}
\description{
A function to estimate a model with parsimoniously time varying parameters.
}
\examples{
\dontrun{
data(economics, package="ggplot2")
psr <- tail(economics$psavert,-1) # personal saving rate.
lur <- head(log(economics$unemploy/economics$pop),-1) # lagged log unemployment rate.
pce <- head(log(economics$pce/economics$pop),-1) # lagged log consumption expenditures per capita.
ptvmod <- ptvfit(psr~lur+pce)
}
}

